--- tauri/src/types/alchemy_response.rs
@@ -1,16 +1,0 @@
-use ethers_core::types::{Address, U256};
-use serde::Deserialize;
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct AlchemyResponse {
-    pub address: Address,
-    pub token_balances: Vec<TokenBalance>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct TokenBalance {
-    pub contract_address: Address,
-    pub token_balance: U256,
-}
