+++ tauri/src/types/mod.rs
@@ -1,9 +1,7 @@
-mod alchemy_response;
 mod checksummed_address;
 pub mod events;
 mod global_state;
 
-pub use alchemy_response::{AlchemyResponse, TokenBalance};
 pub use checksummed_address::ChecksummedAddress;
 pub use events::{Event, Events};
 pub use global_state::GlobalState;
